#108: Split off MapGuide-specific OpenLayers functionality to a separate MapViewerContext class. This class encaspulates all OL classes related to the current runtime map and insulates the ability to switch maps from the map viewer component.
